Start Refactoring
