+++ src/main/kotlin/com/hrblizz/fileapi/controller/FileController.kt
@@ -1,8 +1,8 @@
 package com.hrblizz.fileapi.controller
 
-import com.hrblizz.fileapi.data.entities.Entity
-import com.hrblizz.fileapi.data.repository.EntityRepository
-import com.hrblizz.fileapi.rest.ResponseEntity
+import com.hrblizz.fileapi.data.entities.File
+import com.hrblizz.fileapi.data.repository.FileRepository
+import com.hrblizz.fileapi.dto.FileResponse
 import org.springframework.http.HttpStatus
 import org.springframework.web.bind.annotation.RequestMapping
 import org.springframework.web.bind.annotation.RequestMethod
@@ -10,19 +10,20 @@ import org.springframework.web.bind.annotation.RestController
 import java.util.UUID
 
 @RestController
-class StatusController(
-    private val entityRepository: EntityRepository,
+class FileController(
+    private val entityRepository: FileRepository,
 ) {
+    // TODO getStatus should not Save data by SOLID principles.
     @RequestMapping("/status", method = [RequestMethod.GET])
-    fun getStatus(): ResponseEntity<Map<String, Any>> {
+    fun getStatus(): FileResponse<Map<String, Any>> {
         entityRepository.save(
-            Entity().also {
+            File().also {
                 it.name = UUID.randomUUID().toString()
                 it.value = "asd"
             },
         )
 
-        return ResponseEntity(
+        return FileResponse(
             mapOf(
                 "ok" to true,
             ),
