+++ src/main/kotlin/com/hrblizz/fileapi/data/entities/File.kt
@@ -2,7 +2,8 @@ package com.hrblizz.fileapi.data.entities
 
 import org.springframework.data.annotation.Id
 
-class Entity {
+class File {
+    // TODO name is not ID
     @Id
     lateinit var name: String
     lateinit var value: String
